+++ utils.js
@@ -15,6 +15,13 @@ export function clientPosToCanvasCoords(canvas, clientX, clientY, zoom) {
     return { x, y };
 }
 
+export function clientPosToWorkspaceCoords(workspace, clientX, clientY, zoom) {
+    const rect = workspace.getCanvasBoundingClientRect();
+    const x = Math.floor((clientX - rect.left) / zoom);
+    const y = Math.floor((clientY - rect.top) / zoom);
+    return { x, y };
+}
+
 export function sixBitHexTo0xColor(hex) {
     hex = hex.replace('#', '');
     return ((parseInt(hex, 16) << 8) | 0xFF) >>> 0;