+++ workspace/workspace.js
@@ -1,24 +1,25 @@
 import {Layer} from "./layer.js";
 import {TRANSPARENT_SENTINEL} from "../const.js";
 
-export class Workspace {
-    constructor(width, height, backgroundCanvasEl, foregroundCanvasEl, bgColor = TRANSPARENT_SENTINEL, renderOnCreation = true) {
+class Workspace {
+    constructor() {
+        this.width = null;
+        this.height = null;
+        this.layers = [];
+    }
 
+    init(width, height, canvasElements) {
         this.width = width;
         this.height = height;
-        this.activeLayerIdx = 1;
 
-        this.layers = [
-            new Layer(width, height, backgroundCanvasEl, bgColor, true),
-            new Layer(width, height, foregroundCanvasEl)
-        ];
+        canvasElements.forEach((canvasElement, index) => {
+            const layer = this.addLayer(canvasElement);
+            layer.setZIndex(index + 1);
+            layer.renderInitial();
+        });
 
-        if (renderOnCreation) {
-            this.layers.forEach((layer, index) => {
-                layer.setZIndex(index + 1);
-                layer.renderInitial();
-            });
-        }
+        this.activeLayerIdx = this.layers.length - 1;
+        console.log('workspace initialized');
     }
 
     addLayer(canvasElement) {
@@ -31,12 +32,38 @@ export class Workspace {
         return newLayer;
     }
 
+    removeLayer(idx) {
+        if (idx < 0 || idx >= this.layers.length) return;
+
+        // If removing last layer, clear instead
+        if (this.layers.length === 1) {
+            this.layers[0].clear();
+            return;
+        }
+
+        this.layers[idx].canvas.remove();
+        this.layers.splice(idx, 1);
+
+        // Keep index in place when removing layers
+        if (this.activeLayerIdx === idx) {
+            this.activeLayerIdx = Math.min(this.layers.length - 1, this.activeLayerIdx);
+        } else if (idx < this.activeLayerIdx) {
+            this.activeLayerIdx = Math.max(0, this.activeLayerIdx - 1);
+        } else {
+            this.activeLayerIdx = Math.min(this.layers.length - 1, this.activeLayerIdx);
+        }
+    }
+
     setActiveLayer(idx) {
         this.activeLayerIdx = idx;
     }
 
     getActiveLayer() {
         return this.layers[this.activeLayerIdx];
+    }
+
+    getCanvasBoundingClientRect() {
+        return this.layers[0].canvas.getBoundingClientRect();
     }
 
     update() {
@@ -53,3 +80,5 @@ export class Workspace {
         })
     }
 }
+
+export const workspace = new Workspace();